Skip Archetype's validation when there's only 1 empty fieldset
Previously, marking any properties in an Archetype as required means that the property is effectively required in Umbraco as it forces the user to fill out the initial fieldset.  This means it was impossible to have an Optional property in Umbraco while requiring a field in Archetype when a user DOES add data.

This "fixes"(?) by skipping the validation checks when there's only one fieldset (ie, the initial one) and all of its properties are "empty"

This does remove the ability to require at least 1 fieldset in an Archetype.  As a solution, we could either add a Min Fieldsets prevalue, or we can key off whether the property is marked required in Umbraco (U4-4432)