Replace dependency on Umbraco prevalue
As described in issue #34 (imulus#34) the Archetype configuration
quickly exceeds the 2500 characters allowed by Umbraco prevalues.
This branch is for implementing an attempt to move the Archetype
configuration to a separate custom table within the Umbraco database and
solely use the Umbraco prevalue as an identifier to find the Archetype
configuration in the new table.